Move BlockModel to block package and fix affected imports
