+++ src/main/java/com/notably/model/Model.java
@@ -4,6 +4,7 @@ import java.nio.file.Path;
 import java.util.function.Predicate;
 
 import com.notably.commons.core.GuiSettings;
+import com.notably.model.block.BlockModel;
 import com.notably.model.commandinput.CommandInputModel;
 import com.notably.model.suggestion.SuggestionModel;
